+++ src/com/piposaude/calenjars/types/ddmmm.clj
@@ -5,7 +5,7 @@
             [tick.alpha.api :as t])
   (:import (java.time.format DateTimeParseException)))
 
-(defn get-holiday-ddmm [year name [day month] observed start-year end-year]
+(defn get-holiday-ddmm [year name [day month] observation-rule start-year end-year]
   (try
     (let [holiday (common/holiday name day month year)]
       (cond
@@ -15,11 +15,14 @@
         (and end-year (> year end-year))
         nil
 
-        (and observed (= t/SATURDAY (t/day-of-week (:date holiday))))
+        (and (= observation-rule :observed) (= t/SATURDAY (t/day-of-week (:date holiday))))
         (update holiday :date #(t/- % (t/new-period 1 :days)))
 
-        (and observed (= t/SUNDAY (t/day-of-week (:date holiday))))
+        (and (= observation-rule :observed) (= t/SUNDAY (t/day-of-week (:date holiday))))
         (update holiday :date #(t/+ % (t/new-period 1 :days)))
+
+        (and (= observation-rule :observed-monday-tuesday) (contains? #{t/SATURDAY t/SUNDAY} (t/day-of-week (:date holiday))))
+        (update holiday :date #(t/+ % (t/new-period 2 :days)))
 
         :else
         holiday))