+++ test/com/piposaude/calenjars/holidays_expression_easter_test.clj
@@ -32,6 +32,17 @@
     [{:name "Easter On Monday" :date (t/date "2017-04-17")}] 2017
     [{:name "Easter On Monday" :date (t/date "2018-04-02")}] 2018))
 
+(deftest should-generate-holidays-when-holidays-for-year-with-easter-expression-and-monday-tuesday-observance-rule
+  (are [expected year]
+       (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-observed-monday-tuesday.hol"))
+    [{:name "Easter On Tuesday" :date (t/date "2012-04-10")}] 2012
+    [{:name "Easter On Tuesday" :date (t/date "2013-04-02")}] 2013
+    [{:name "Easter On Tuesday" :date (t/date "2014-04-22")}] 2014
+    [{:name "Easter On Tuesday" :date (t/date "2015-04-07")}] 2015
+    [{:name "Easter On Tuesday" :date (t/date "2016-03-29")}] 2016
+    [{:name "Easter On Tuesday" :date (t/date "2017-04-18")}] 2017
+    [{:name "Easter On Tuesday" :date (t/date "2018-04-03")}] 2018))
+
 (deftest should-generate-holidays-when-holidays-for-year-with-easter-expression-and-start-clause
   (are [expected year]
        (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-start.hol"))
@@ -75,6 +86,27 @@
     [{:name "Easter On Monday" :date (t/date "2017-04-17")}] 2017
     [{:name "Easter On Monday" :date (t/date "2018-04-02")}] 2018))
 
+(deftest should-generate-holidays-when-holidays-for-year-with-easter-expression-with-monday-tuesday-observance-rule-and-start-clause
+  (are [expected year]
+       (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-observed-monday-tuesday-start.hol"))
+    [] 2012
+    [] 2013
+    [] 2014
+    [{:name "Easter On Tuesday" :date (t/date "2015-04-07")}] 2015
+    [{:name "Easter On Tuesday" :date (t/date "2016-03-29")}] 2016
+    [{:name "Easter On Tuesday" :date (t/date "2017-04-18")}] 2017
+    [{:name "Easter On Tuesday" :date (t/date "2018-04-03")}] 2018)
+
+  (are [expected year]
+       (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-observed-monday-tuesday-start-reverse-order.hol"))
+    [] 2012
+    [] 2013
+    [] 2014
+    [{:name "Easter On Tuesday" :date (t/date "2015-04-07")}] 2015
+    [{:name "Easter On Tuesday" :date (t/date "2016-03-29")}] 2016
+    [{:name "Easter On Tuesday" :date (t/date "2017-04-18")}] 2017
+    [{:name "Easter On Tuesday" :date (t/date "2018-04-03")}] 2018))
+
 (deftest should-generate-holidays-when-holidays-for-year-with-easter-expression-and-end-clause
   (are [expected year]
        (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-end.hol"))
@@ -115,5 +147,26 @@
     [{:name "Easter On Monday" :date (t/date "2014-04-21")}] 2014
     [{:name "Easter On Monday" :date (t/date "2015-04-06")}] 2015
     [{:name "Easter On Monday" :date (t/date "2016-03-28")}] 2016
+    [] 2017
+    [] 2018))
+
+(deftest should-generate-holidays-when-holidays-for-year-with-easter-expression-with-monday-tuesday-observance-rule-and-end-clause
+  (are [expected year]
+       (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-observed-monday-tuesday-end.hol"))
+    [{:name "Easter On Tuesday" :date (t/date "2012-04-10")}] 2012
+    [{:name "Easter On Tuesday" :date (t/date "2013-04-02")}] 2013
+    [{:name "Easter On Tuesday" :date (t/date "2014-04-22")}] 2014
+    [{:name "Easter On Tuesday" :date (t/date "2015-04-07")}] 2015
+    [{:name "Easter On Tuesday" :date (t/date "2016-03-29")}] 2016
+    [] 2017
+    [] 2018)
+
+  (are [expected year]
+       (= expected (gen/holidays-for-year year "test-resources/generate/expression-easter-observed-monday-tuesday-end-reverse-order.hol"))
+    [{:name "Easter On Tuesday" :date (t/date "2012-04-10")}] 2012
+    [{:name "Easter On Tuesday" :date (t/date "2013-04-02")}] 2013
+    [{:name "Easter On Tuesday" :date (t/date "2014-04-22")}] 2014
+    [{:name "Easter On Tuesday" :date (t/date "2015-04-07")}] 2015
+    [{:name "Easter On Tuesday" :date (t/date "2016-03-29")}] 2016
     [] 2017
     [] 2018))